chore: extract rive logic into hook
