+++ app/components/UI/Bridge/hooks/useRewardsIconAnimation/index.ts
@@ -1,0 +1,1 @@
+export { useRewardsIconAnimation } from './useRewardsIconAnimation';
